requirements.txt
certifi==2020.12.5
elasticsearch==7.12.0
numpy==1.17.2
pandas==1.2.3
pymongo==3.11.3
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
urllib3==1.26.4
requests_aws4auth
python-decouple
s3fs